refactor(monorepo): remove hls-utils, move utilities into consumers (#25)
- Move resolveUrl into hls-base (src/url.ts), update pipeline and vitest config
- Move parseAttributes, parseResolution, parseByteRange into hls-parser (src/attributes.ts)
- Add unit tests for moved code in each package
- Remove @mtngtools/hls-utils from deps, root tsconfig, and docs (README, spec)
- Delete packages/hls-utils (unused exports dropped)

Closes #25
