+++ packages/hls-parser/package.json
@@ -25,8 +25,7 @@
     "test:e2e": "E2E_ENABLED=true vitest run"
   },
   "dependencies": {
-    "@mtngtools/hls-types": "workspace:*",
-    "@mtngtools/hls-utils": "workspace:*"
+    "@mtngtools/hls-types": "workspace:*"
   },
   "keywords": [
     "hls",
