--- packages/hls-utils/README.md
@@ -1,34 +1,0 @@
-# @mtngtools/hls-utils
-
-Common utilities for HLS operations.
-
-## Overview
-
-Shared utility functions used across HLS packages.
-
-## Installation
-
-```bash
-pnpm add @mtngtools/hls-utils
-```
-
-## Usage
-
-```typescript
-import { /* utilities */ } from '@mtngtools/hls-utils';
-```
-
-## Features
-
-- Minimal dependencies
-- Runtime-agnostic utilities
-- Common helper functions
-
-## Specifications
-  
-See [Specification Details](./spec/README.md) for detailed specifications.
-
-## License
-
-MIT
-
