--- packages/hls-utils/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@mtngtools/hls-utils",
-  "version": "0.1.0",
-  "description": "Common utilities for HLS operations",
-  "type": "module",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "vite build",
-    "dev": "vite build --watch",
-    "clean": "rimraf dist",
-    "typecheck": "tsc --noEmit",
-    "lint": "oxlint",
-    "test": "vitest run",
-    "test:watch": "vitest",
-    "test:e2e": "E2E_ENABLED=true vitest run"
-  },
-  "dependencies": {
-    "@mtngtools/hls-types": "workspace:*"
-  },
-  "keywords": [
-    "hls",
-    "utilities",
-    "helpers"
-  ],
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  }
-}
-
