--- packages/hls-utils/vite.config.ts
@@ -1,23 +1,0 @@
-import { defineConfig } from 'vite';
-import { resolve } from 'path';
-import dts from 'vite-plugin-dts';
-
-export default defineConfig({
-  plugins: [
-    dts({
-      insertTypesEntry: true,
-    }),
-  ],
-  build: {
-    lib: {
-      entry: resolve(__dirname, 'src/index.ts'),
-      formats: ['es'],
-      fileName: 'index',
-    },
-    rollupOptions: {
-      external: ['@mtngtools/hls-types'],
-    },
-    outDir: 'dist',
-  },
-});
-
